# Speedup `character_length`