## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

- Closes #15930

## Rationale for this change



character_length benchmark
<details>

```
character_length_StringArray_ascii_str_len_8
                        time:   [9.2558 µs 9.2759 µs 9.2971 µs]
                        change: [-69.094% -69.031% -68.961%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

character_length_StringArray_utf8_str_len_8
                        time:   [51.115 µs 51.980 µs 52.962 µs]
                        change: [-50.042% -49.011% -47.983%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  15 (15.00%) high mild

character_length_StringViewArray_ascii_str_len_8
                        time:   [21.520 µs 21.669 µs 21.848 µs]
                        change: [-49.637% -49.231% -48.906%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

character_length_StringViewArray_utf8_str_len_8
                        time:   [63.238 µs 64.204 µs 65.335 µs]
                        change: [-49.047% -47.810% -46.388%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

character_length_StringArray_ascii_str_len_32
                        time:   [16.314 µs 16.388 µs 16.471 µs]
                        change: [-55.729% -55.606% -55.480%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

character_length_StringArray_utf8_str_len_32
                        time:   [85.483 µs 86.380 µs 87.239 µs]
                        change: [-27.927% -27.184% -26.451%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

character_length_StringViewArray_ascii_str_len_32
                        time:   [30.754 µs 30.849 µs 30.958 µs]
                        change: [-40.686% -40.498% -40.322%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

character_length_StringViewArray_utf8_str_len_32
                        time:   [120.58 µs 121.98 µs 123.22 µs]
                        change: [-22.085% -21.127% -20.191%] (p = 0.00 < 0.05)
                        Performance has improved.

character_length_StringArray_ascii_str_len_128
                        time:   [44.115 µs 44.194 µs 44.347 µs]
                        change: [-31.392% -31.296% -31.166%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

character_length_StringArray_utf8_str_len_128
                        time:   [138.57 µs 138.89 µs 139.12 µs]
                        change: [-13.267% -12.873% -12.475%] (p = 0.00 < 0.05)
                        Performance has improved.

character_length_StringViewArray_ascii_str_len_128
                        time:   [62.667 µs 62.775 µs 62.959 µs]
                        change: [-26.719% -26.554% -26.404%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  6 (6.00%) high severe

character_length_StringViewArray_utf8_str_len_128
                        time:   [166.17 µs 166.76 µs 167.41 µs]
                        change: [-14.128% -13.824% -13.512%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  12 (12.00%) high mild
  8 (8.00%) high severe

Benchmarking character_length_StringArray_ascii_str_len_4096: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.4s, enable flat sampling, or reduce sample count to 60.
character_length_StringArray_ascii_str_len_4096
                        time:   [1.2724 ms 1.2755 ms 1.2791 ms]
                        change: [-1.6305% -1.2835% -0.9450%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) low severe
  4 (4.00%) low mild
  5 (5.00%) high severe

character_length_StringArray_utf8_str_len_4096
                        time:   [2.1568 ms 2.1614 ms 2.1654 ms]
                        change: [-0.2057% +0.0679% +0.3527%] (p = 0.65 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low severe
  3 (3.00%) low mild

Benchmarking character_length_StringViewArray_ascii_str_len_4096: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.4s, enable flat sampling, or reduce sample count to 50.
character_length_StringViewArray_ascii_str_len_4096
                        time:   [1.4566 ms 1.4609 ms 1.4647 ms]
                        change: [-1.6508% -1.3910% -1.1291%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  3 (3.00%) low severe
  5 (5.00%) low mild
  2 (2.00%) high mild
  5 (5.00%) high severe

character_length_StringViewArray_utf8_str_len_4096
                        time:   [2.2177 ms 2.2231 ms 2.2284 ms]
                        change: [+0.5513% +0.8467% +1.1138%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) low mild
```
</details>
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

## What changes are included in this PR?

Closes: #15930

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->