@Dandandan 2025-05-02
The main speedup I am expecting in e2e benchmarks is query 27 of clickbench, which has some mixed ascii / utf8 data and uses a `LENGTH` function.
Local runs don't show a very large diff (around 6%) - profiling shows a larger diff from ~15% to ~3% of samples in do_count_chars (i.e. `.chars()

A large part from the speedup in the micro benchmark seems to come the faster array creation (collect into Vec instead of PrimitiveBuilder).