@alamb 2025-05-03
🤖 `./gh_compare_branch_bench.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_branch_bench.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing speedup_character_length (3ee71fe401848bdd84666766f157dec91aef4cda) to d2a2a8be7cfd6ab479f4c231d78ff2481cd79209 [diff](https://github.com/apache/datafusion/compare/d2a2a8be7cfd6ab479f4c231d78ff2481cd79209..3ee71fe401848bdd84666766f157dec91aef4cda)
BENCH_NAME=character_length
BENCH_COMMAND=cargo bench --bench character_length
BENCH_FILTER=
BENCH_BRANCH_NAME=speedup_character_length
Results will be posted here when complete