@alamb 2025-05-03
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                  main                                   speedup_character_length
-----                                                  ----                                   ------------------------
character_length_StringArray_ascii_str_len_128         1.97     76.7±1.10µs        ? ?/sec    1.00     38.9±1.18µs        ? ?/sec
character_length_StringArray_ascii_str_len_32          2.85     57.6±0.16µs        ? ?/sec    1.00     20.2±0.07µs        ? ?/sec
character_length_StringArray_ascii_str_len_4096        1.04      2.2±0.10ms        ? ?/sec    1.00      2.1±0.04ms        ? ?/sec
character_length_StringArray_ascii_str_len_8           3.27     54.1±0.18µs        ? ?/sec    1.00     16.6±0.09µs        ? ?/sec
character_length_StringArray_utf8_str_len_128          1.16    287.0±0.71µs        ? ?/sec    1.00    246.7±1.19µs        ? ?/sec
character_length_StringArray_utf8_str_len_32           1.20    222.4±0.35µs        ? ?/sec    1.00    186.0±1.18µs        ? ?/sec
character_length_StringArray_utf8_str_len_4096         1.19      6.3±0.12ms        ? ?/sec    1.00      5.3±0.07ms        ? ?/sec
character_length_StringArray_utf8_str_len_8            1.06    157.0±0.33µs        ? ?/sec    1.00    148.6±0.23µs        ? ?/sec
character_length_StringViewArray_ascii_str_len_128     1.31    105.0±0.78µs        ? ?/sec    1.00     80.2±0.23µs        ? ?/sec
character_length_StringViewArray_ascii_str_len_32      1.52     90.7±0.14µs        ? ?/sec    1.00     59.7±0.73µs        ? ?/sec
character_length_StringViewArray_ascii_str_len_4096    1.04      2.1±0.07ms        ? ?/sec    1.00  1995.4±55.05µs        ? ?/sec
character_length_StringViewArray_ascii_str_len_8       1.68    105.3±0.11µs        ? ?/sec    1.00     62.5±0.17µs        ? ?/sec
character_length_StringViewArray_utf8_str_len_128      1.17    301.8±1.07µs        ? ?/sec    1.00    257.5±0.66µs        ? ?/sec
character_length_StringViewArray_utf8_str_len_32       1.19    237.7±0.62µs        ? ?/sec    1.00    199.0±0.51µs        ? ?/sec
character_length_StringViewArray_utf8_str_len_4096     1.22      6.4±0.11ms        ? ?/sec    1.00      5.3±0.10ms        ? ?/sec
character_length_StringViewArray_utf8_str_len_8        1.02    167.2±0.32µs        ? ?/sec    1.00    164.2±0.29µs        ? ?/sec
```


</p>
</details>