@alamb 2025-05-05
🤖 `./gh_compare_branch.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_branch.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing speedup_character_length (2a789fb4ab0029ff3f4f3d9d1e0fe5354ebd5c7b) to b4b77e949d9f1c11242a6f34d8eb27f18ad4d807 [diff](https://github.com/apache/datafusion/compare/b4b77e949d9f1c11242a6f34d8eb27f18ad4d807..2a789fb4ab0029ff3f4f3d9d1e0fe5354ebd5c7b)
Benchmarks: tpch_mem clickbench_partitioned clickbench_extended
Results will be posted here when complete